requirements.txt
streamlit
PyMuPDF
spacy
python-docx
sentence-transformers
torch
